Fix (most of) everything needed for upstreaming
